Bug 1139020: Update SCTP library from upstream to rev 9168 rs=jesup
UltraBlame original commit: 01f4c9a6e84048db839265ccb44006f5e4f94a3d

+++ netwerk/sctp/src/user_route.h
@@ -54,7 +54,7 @@ struct sctp_route {
 
 
 struct sctp_rt_metrics_lite {
-	u_long	rmx_mtu;	
+	uint32_t rmx_mtu;	
 #if 0
 	u_long	rmx_expire;	
 	u_long	rmx_pksent;
